feat(libft, builtins): Implementa ft_atoul e ft_is_long; aprimora Makefile e builtin export

+++ libft/srcs/stdlib/ft_atoul.c
@@ -26,7 +26,7 @@ unsigned long	ft_atoul(char *nptr)
 	while (nptr[i] >= '0' && nptr[i] <= '9')
 	{
 		if (result > ULONG_MAX / 10 || (result == ULONG_MAX / 10
-				&& (nptr[i] - '0') > ULONG_MAX % 10))
+				&& (unsigned long)(nptr[i] - '0') > ULONG_MAX % 10))
 			return (ULONG_MAX);
 		result = (result * 10) + (nptr[i++] - '0');
 	}
